Rename base class for strict dependencies integration tests
This class didn't have anything specific to do with strict dependencies. It has been renamed
and moved to the integration test fixtures, so that it can be reused by other tests (in prevision
for increasing coverage of Gradle metadata).
